+++ .githooks/bash/uninstall.sh
