+++ .githooks/symlinks.js
