+++ package.json
@@ -113,8 +113,11 @@
         "sw-precache": "5.2.0"
     },
     "scripts": {
+        "create-hooks": "node .githooks/create-hooks.js",
+        "remove-hooks": "node .githooks/remove-hooks.js",
         "lint": "npm run lint:javascript && polymer lint",
         "lint:javascript": "eslint . --ext js,html --ignore-path .gitignore",
-        "build": "gulp build:es5 && gulp build:es6"
+        "build": "gulp build:es5 && gulp build:es6",
+        "start": "npm run create-hooks"
     }
 }
